split logic in conversations panel into hooks
