+++ front-react/src/getWrapStatus.ts
@@ -1,4 +1,5 @@
 export const WRAP_STORAGE_KEY = 'conversation-wrap-enabled';
+export const WRAP_STATUS_EVENT = 'conversation-wrap-change';
 
 export function getWrapStatus(): boolean {
   return localStorage.getItem(WRAP_STORAGE_KEY) === 'true';
@@ -6,4 +7,8 @@ export function getWrapStatus(): boolean {
 
 export function setWrapStatus(newValue: boolean) {
   localStorage.setItem(WRAP_STORAGE_KEY, String(newValue));
+
+  if (typeof window !== 'undefined') {
+    window.dispatchEvent(new CustomEvent<boolean>(WRAP_STATUS_EVENT, { detail: newValue }));
+  }
 }
